+++ packages/api/src/api.test.ts
@@ -103,6 +103,26 @@ describe('api', () => {
         //assert
         expect(response.status).toBe(500)
       })
+
+      it('throw 400 error if invalid groupBy is passed', async () => {
+        //setup
+        const startDate = '2020-07-12'
+        const endDate = '2020-07-13'
+        const groupBy = 'hello'
+
+        const expectedResponse: EstimationResult[] = []
+        mockGetCostAndEstimates.mockResolvedValueOnce(expectedResponse)
+
+        //run
+        const response = await request(server).get(
+          encodeURI(
+            `/footprint?start=${startDate}&end=${endDate}&groupBy=${groupBy}`,
+          ),
+        )
+
+        //assert
+        expect(response.status).toBe(400)
+      })
     })
   })