+++ packages/api/src/api.ts
@@ -38,10 +38,15 @@ const FootprintApiMiddleware = async function (
     startDate: req.query.start?.toString(),
     endDate: req.query.end?.toString(),
     ignoreCache: req.query.ignoreCache?.toString(),
+    groupBy: req.query.groupBy?.toString(),
   }
   apiLogger.info(
     `Footprint API request started with Start Date: ${rawRequest.startDate} and End Date: ${rawRequest.endDate}`,
   )
+  if (!rawRequest.groupBy)
+    apiLogger.warn(
+      'GroupBy parameter not specified. This will be required in the future.',
+    )
   const footprintApp = new App()
   try {
     const estimationRequest = CreateValidFootprintRequest(rawRequest)