+++ packages/app/src/__tests__/Cache.test.ts
@@ -3,15 +3,15 @@
  */
 
 import moment from 'moment'
-import DurationConstructor = moment.unitOfTime.DurationConstructor
 import {
-  configLoader,
   EstimationResult,
+  GroupBy,
   ServiceData,
 } from '@cloud-carbon-footprint/common'
 import cache from '../Cache'
 import { EstimationRequest } from '../CreateValidRequest'
 import CacheManager from '../CacheManager'
+import DurationConstructor = moment.unitOfTime.DurationConstructor
 
 let mockSetEstimates: jest.Mock
 let mockGetEstimates: jest.Mock
@@ -26,18 +26,6 @@ jest.mock('../CacheManager', () => {
     }
   })
 })
-
-jest.mock('@cloud-carbon-footprint/common', () => ({
-  ...(jest.requireActual('@cloud-carbon-footprint/common') as Record<
-    string,
-    unknown
-  >),
-  configLoader: jest.fn().mockImplementation(() => {
-    return {
-      GROUP_QUERY_RESULTS_BY: 'day',
-    }
-  }),
-}))
 
 const dummyServiceEstimate: ServiceData[] = [
   {
@@ -84,14 +72,6 @@ describe('Cache', () => {
   })
 
   describe('cache-returned function per day', () => {
-    beforeEach(() => {
-      ;(configLoader as jest.Mock).mockReturnValue({
-        GROUP_QUERY_RESULTS_BY: 'day',
-      })
-    })
-    afterEach(() => {
-      jest.restoreAllMocks()
-    })
     it('returns cached data from cache service instead of calling the real method', async () => {
       //setup
       const rawRequest: EstimationRequest = {
@@ -101,7 +81,7 @@ describe('Cache', () => {
       }
 
       const expectedEstimationResults: EstimationResult[] =
-        buildFootprintEstimates('2020-01-01', 1, dummyServiceEstimate)
+        buildFootprintEstimates('2020-01-01', 2, dummyServiceEstimate)
       mockGetEstimates.mockResolvedValueOnce(expectedEstimationResults)
 
       const target = {}
@@ -187,7 +167,37 @@ describe('Cache', () => {
         startDate: moment.utc('2020-01-01').toDate(),
         endDate: moment.utc('2020-01-02').toDate(),
         ignoreCache: false,
+        groupBy: undefined,
+        region: 'us-east-1',
       })
+    })
+
+    it('calls original function with grouping in the request', async () => {
+      //setup
+      const rawRequest: EstimationRequest = {
+        startDate: moment.utc('2019-12-31').toDate(),
+        endDate: moment.utc('2020-01-02').toDate(),
+        region: 'us-east-1',
+        ignoreCache: false,
+        groupBy: GroupBy.week,
+      }
+
+      const cachedEstimates: EstimationResult[] = buildFootprintEstimates(
+        '2019-12-31',
+        7,
+      )
+
+      mockGetEstimates.mockResolvedValueOnce(cachedEstimates)
+
+      //run
+      cacheDecorator({}, 'propertyTest', propertyDescriptor)
+      await propertyDescriptor.value(rawRequest)
+
+      //assert
+      expect(mockGetEstimates).toHaveBeenCalledWith(
+        rawRequest,
+        rawRequest.groupBy,
+      )
     })
 
     it('does not fetch dates when cache service returns unordered estimates', async () => {
@@ -234,7 +244,7 @@ describe('Cache', () => {
 
       mockGetEstimates.mockResolvedValueOnce(cachedEstimates)
 
-      const computedEstimates = buildFootprintEstimates('2019-12-31', 1)
+      const computedEstimates = buildFootprintEstimates('2019-12-31', 2)
       originalFunction.mockResolvedValueOnce(computedEstimates)
 
       //run
@@ -309,7 +319,7 @@ describe('Cache', () => {
 
       //assert
       expect(mockSetEstimates).toHaveBeenCalledWith(
-        computedEstimates.concat(buildFootprintEstimates('2020-07-15', 5)),
+        computedEstimates.concat(buildFootprintEstimates('2020-07-15', 6)),
         'day',
       )
     })
@@ -324,7 +334,7 @@ describe('Cache', () => {
 
       const cachedEstimates: EstimationResult[] = buildFootprintEstimates(
         '2020-07-15',
-        5,
+        6,
       )
       mockGetEstimates.mockResolvedValueOnce(cachedEstimates)
 
@@ -353,66 +363,6 @@ describe('Cache', () => {
 
       //assert
       expect(results).toEqual(computedEstimates)
-    })
-  })
-
-  describe('cache-returned function per week', () => {
-    beforeEach(() => {
-      ;(configLoader as jest.Mock).mockReturnValue({
-        GROUP_QUERY_RESULTS_BY: 'week',
-      })
-    })
-    afterEach(() => {
-      jest.restoreAllMocks()
-    })
-    it('fetches dates not stored in cache', async () => {
-      //given
-      const rawRequest: EstimationRequest = {
-        startDate: moment.utc('2021-01-01').toDate(),
-        endDate: moment.utc('2021-02-01').toDate(),
-        ignoreCache: false,
-      }
-
-      const cachedEstimates: EstimationResult[] = buildFootprintEstimates(
-        '2021-01-01',
-        2,
-        dummyServiceEstimate,
-        'weeks',
-      )
-
-      mockGetEstimates.mockResolvedValueOnce(cachedEstimates)
-
-      const computedEstimates1 = buildFootprintEstimates(
-        '2021-01-15',
-        2,
-        dummyServiceEstimate,
-        'weeks',
-      )
-
-      const computedEstimates2 = buildFootprintEstimates(
-        '2021-01-29',
-        2,
-        dummyServiceEstimate,
-        'weeks',
-      )
-
-      originalFunction
-        .mockResolvedValueOnce(computedEstimates1)
-        .mockResolvedValueOnce(computedEstimates2)
-
-      //when
-      cacheDecorator({}, 'propertyTest', propertyDescriptor)
-      const estimationResult: EstimationResult[] =
-        await propertyDescriptor.value(rawRequest)
-
-      //then
-      const expectedEstimationResults: EstimationResult[] = [
-        ...cachedEstimates,
-        ...computedEstimates1,
-        ...computedEstimates2,
-      ]
-
-      expect(estimationResult).toEqual(expectedEstimationResults)
     })
   })
 })