+++ packages/app/src/__tests__/CacheManager.test.ts
@@ -6,9 +6,11 @@ import CacheManager from '../CacheManager'
 import EstimatorCacheGoogleCloudStorage from '../EstimatorCacheGoogleCloudStorage'
 import EstimatorCacheFileSystem from '../EstimatorCacheFileSystem'
 import { EstimationRequest } from '../CreateValidRequest'
-import { EstimationResult } from '@cloud-carbon-footprint/common'
-
-import { Config as mockConfig } from '@cloud-carbon-footprint/common'
+import {
+  Config as mockConfig,
+  EstimationResult,
+  GroupBy,
+} from '@cloud-carbon-footprint/common'
 
 function buildFootprintEstimates(startDate: string, consecutiveDays: number) {
   return [...Array(consecutiveDays)].map((v, i) => {
@@ -19,6 +21,7 @@ function buildFootprintEstimates(startDate: string, consecutiveDays: number) {
   })
 }
 
+const grouping = GroupBy.week
 describe('CacheManager - CACHE_MODE: GCS', () => {
   beforeAll(() => {
     mockConfig.GCP.CACHE_BUCKET_NAME = 'test-bucket-name'
@@ -49,8 +52,12 @@ describe('CacheManager - CACHE_MODE: GCS', () => {
         startDate: moment.utc(startDate).toDate(),
         endDate: moment.utc(endDate).toDate(),
         ignoreCache: false,
+        groupBy: grouping,
       }
-      const estimates = await cacheManager.getEstimates(request)
+      const estimates = await cacheManager.getEstimates(
+        request,
+        request.groupBy,
+      )
 
       //assert
       expect(estimates).toEqual(buildFootprintEstimates(startDate, 1))
@@ -76,11 +83,15 @@ describe('CacheManager - CACHE_MODE: GCS', () => {
         startDate: moment.utc(startDate).toDate(),
         endDate: moment.utc(endDate).toDate(),
         ignoreCache: false,
+        groupBy: grouping,
       }
-      const estimates = await cacheManager.getEstimates(request)
+      const estimates = await cacheManager.getEstimates(
+        request,
+        request.groupBy,
+      )
 
       //assert
-      expect(estimates).toEqual(buildFootprintEstimates(startDate, 2))
+      expect(estimates).toEqual(buildFootprintEstimates(startDate, 3))
     })
   })
 
@@ -104,7 +115,7 @@ describe('CacheManager - CACHE_MODE: GCS', () => {
       const cacheManager = new CacheManager()
 
       //run
-      const estimates = await cacheManager.setEstimates(cachedData)
+      const estimates = await cacheManager.setEstimates(cachedData, 'day')
 
       //assert
       expect(estimates).resolves
@@ -124,7 +135,7 @@ describe('CacheManager - CACHE_MODE: GCS', () => {
       )
 
       //run
-      const estimates = await cacheManager.setEstimates(cachedData)
+      const estimates = await cacheManager.setEstimates(cachedData, 'day')
 
       //assert
       expect(estimates).resolves
@@ -156,8 +167,12 @@ describe('CacheManager - CACHE_MODE: fileSystem', () => {
         startDate: moment.utc(startDate).toDate(),
         endDate: moment.utc(endDate).toDate(),
         ignoreCache: false,
+        groupBy: grouping,
       }
-      const estimates = await cacheManager.getEstimates(request)
+      const estimates = await cacheManager.getEstimates(
+        request,
+        request.groupBy,
+      )
 
       //assert
       expect(estimates).toEqual(buildFootprintEstimates(startDate, 1))
@@ -179,11 +194,15 @@ describe('CacheManager - CACHE_MODE: fileSystem', () => {
         startDate: moment.utc(startDate).toDate(),
         endDate: moment.utc(endDate).toDate(),
         ignoreCache: false,
+        groupBy: grouping,
       }
-      const estimates = await cacheManager.getEstimates(request)
+      const estimates = await cacheManager.getEstimates(
+        request,
+        request.groupBy,
+      )
 
       //assert
-      expect(estimates).toEqual(buildFootprintEstimates(startDate, 2))
+      expect(estimates).toEqual(buildFootprintEstimates(startDate, 3))
     })
   })
 
@@ -207,7 +226,7 @@ describe('CacheManager - CACHE_MODE: fileSystem', () => {
       const cacheManager = new CacheManager()
 
       //run
-      const estimates = await cacheManager.setEstimates(cachedData)
+      const estimates = await cacheManager.setEstimates(cachedData, 'day')
 
       //assert
       expect(estimates).resolves
@@ -228,7 +247,7 @@ describe('CacheManager - CACHE_MODE: fileSystem', () => {
       )
 
       //run
-      const estimates = await cacheManager.setEstimates(cachedData)
+      const estimates = await cacheManager.setEstimates(cachedData, 'day')
 
       //assert
       expect(estimates).resolves